feat(plugins): move ClusterDependencies to keel-ec2-api
