+++ keel-ec2-api/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/LaunchConfiguration.kt
