+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/jackson/mixins/ClusterDependenciesMixin.kt
@@ -1,0 +1,7 @@
+package com.netflix.spinnaker.keel.ec2.jackson.mixins
+
+import com.fasterxml.jackson.annotation.JsonInclude
+import com.fasterxml.jackson.annotation.JsonInclude.Include.NON_EMPTY
+
+@JsonInclude(NON_EMPTY)
+interface ClusterDependenciesMixin
