+++ keel-titus-plugin/keel-titus-plugin.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 
 dependencies {
   api(project(":keel-api"))
+  api(project(":keel-ec2-api"))
   implementation(project(":keel-core")) // TODO: ideally not
   implementation(project(":keel-clouddriver"))
   implementation(project(":keel-orca"))
